fix: support echoctl on Windows
